backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiohttp
pandas
apscheduler
python-dotenv
